+++ sample/src/main/java/no/nordicsemi/kotlin/ble/android/sample/advertiser/AdvertiserScreen.kt
@@ -31,33 +31,84 @@
 
 package no.nordicsemi.kotlin.ble.android.sample.advertiser
 
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import no.nordicsemi.kotlin.ble.core.android.AndroidEnvironment
+import no.nordicsemi.kotlin.ble.environment.android.compose.LocalEnvironmentOwner
 
 @Composable
 fun AdvertiserScreen() {
+    val environment = LocalEnvironmentOwner.current
     val vm = hiltViewModel<AdvertiserViewModel>()
-    val state by vm.isAdvertising.collectAsStateWithLifecycle()
+    val state by environment.bluetoothState.collectAsStateWithLifecycle()
+    val isAdvertising by vm.isAdvertising.collectAsStateWithLifecycle()
     val error by vm.error.collectAsStateWithLifecycle()
 
-    AdvertiserView(
-        isAdvertising = state,
-        onStartClicked = vm::startAdvertising,
-        onStopClicked = vm::stopAdvertising,
-        errorMessage = error,
+    Column(
         modifier = Modifier
             .fillMaxSize()
-            .padding(horizontal = 16.dp)
-            .verticalScroll(rememberScrollState())
-            .padding(top = 16.dp, bottom = 32.dp),
-        environment = vm.environment,
-    )
+            .padding(horizontal = 16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(16.dp),
+    ) {
+        Text(text = "Bluetooth state: $state")
+
+        // Advertising requires BLUETOOTH_ADVERTISE permission.
+        val permissions = arrayOf(
+            AndroidEnvironment.Permission.BLUETOOTH_ADVERTISE,
+        )
+        var permissionGranted by remember {
+            mutableStateOf(environment.isBluetoothAdvertisePermissionGranted)
+        }
+        val launcher = rememberLauncherForActivityResult(
+            contract = ActivityResultContracts.RequestMultiplePermissions(),
+            onResult = {
+                // This may not work.
+                // permissionGranted = it.values.all { true }
+                // Use this instead:
+                permissionGranted = environment.isBluetoothAdvertisePermissionGranted
+            }
+        )
+
+        if (permissionGranted) {
+            // Both Bluetooth and Location permissions are granted.
+            // We can now start scanning.
+            AdvertiserView(
+                isAdvertising = isAdvertising,
+                onStartClicked = vm::startAdvertising,
+                onStopClicked = vm::stopAdvertising,
+                errorMessage = error,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(horizontal = 16.dp)
+                    .verticalScroll(rememberScrollState())
+                    .padding(top = 16.dp, bottom = 32.dp),
+                environment = environment,
+            )
+        } else {
+            Button(
+                onClick = {  launcher.launch(permissions) }
+            ) {
+                Text("Grant required permissions")
+            }
+        }
+    }
 }
