+++ sample/src/main/java/no/nordicsemi/kotlin/ble/android/sample/advertiser/AdvertiserView.kt
@@ -118,7 +118,6 @@ fun AdvertiserView(
             )
 
             AnimatedVisibility(visible = propertiesVisible) {
-
                 Column(
                     modifier = Modifier.padding(vertical = 8.dp),
                     verticalArrangement = Arrangement.spacedBy(8.dp),
