+++ sample/src/main/java/no/nordicsemi/kotlin/ble/android/sample/scanner/ScannerViewModel.kt
@@ -59,6 +59,7 @@ import no.nordicsemi.kotlin.ble.client.android.preview.PreviewPeripheral
 import no.nordicsemi.kotlin.ble.client.distinctByPeripheral
 import no.nordicsemi.kotlin.ble.client.exception.InvalidAttributeException
 import no.nordicsemi.kotlin.ble.core.ConnectionState
+import no.nordicsemi.kotlin.ble.core.IncludedService
 import no.nordicsemi.kotlin.ble.core.Phy
 import no.nordicsemi.kotlin.ble.core.PhyInUse
 import no.nordicsemi.kotlin.ble.core.WriteType
@@ -321,7 +322,7 @@ class ScannerViewModel @Inject constructor(
             .onEach { services ->
                 // On each services change, increment the event index.
                 event += 1
-                Timber.i("($event) Services changed: $services")
+                Timber.i("($event) Services changed: ${services?.map { it.uuid to it.isPrimary }}")
             }
             .filterNotNull()
             .onEach { services ->
@@ -393,7 +394,9 @@ class ScannerViewModel @Inject constructor(
                 // Check if LED Button service is available.
                 // If so, blink the LED 5 times.
                 val blinkyServiceUuid = Uuid.parse("00001523-1212-efde-1523-785feabcd123")
-                val blinkyService = services.firstOrNull { it.uuid == blinkyServiceUuid }
+                val blinkyService = services.firstOrNull {
+                    it.uuid == blinkyServiceUuid && it.isPrimary
+                }
                 blinkyService?.let { service ->
                     val buttonCharacteristicUuid = Uuid.parse("00001524-1212-efde-1523-785feabcd123")
                     val ledCharacteristicUuid = Uuid.parse("00001525-1212-efde-1523-785feabcd123")
